requirements.txt
flask==3.0.3
python-dotenv==1.0.1
pyodbc==5.1.0
pymssql==2.3.1
PyJWT==2.8.0
bcrypt==4.2.0
cloudinary==1.40.0
requests==2.32.3
Werkzeug==3.0.3
gunicorn==22.0.0
Pillow==10.4.0
